+++ CHANGELOG.md
@@ -128,3 +128,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - fix destroying issues
 - add option onPaymentStarted to initMethod
+
+## [0.5.3] - 2025-01-13
+
+- default skin clear own container
+
+## [0.5.4] - 2025-01-16
+
+- iniMethod create headless checkouts in order
+- clientSession cache fix
+- improve examples
+- add option onAvaialbleMethods to initMethod
+
+## [0.5.5] - 2025-01-19
+
+- fix cardholder name input
+- make initMethod callbacks optional